+++ core/math/src/main/java/org/arakhne/afc/math/geometry/d3/ai/PathShadow3ai.java
@@ -302,7 +302,7 @@ public class PathShadow3ai<B extends RectangularPrism3ai<?, ?, ?, ?, ?, B>> {
         if (sx0 > shadowXmax && sx1 > shadowXmax) {
 			// The line is entirely at the right of the shadow
 			if (sy1 != sy0) {
-                final double alpha = (sx1 - sx0) / (sy1 - sy0);
+                final double alpha = (double) (sx1 - sx0) / (sy1 - sy0);
                 if (sy0 < sy1) {
                     if (sy0 <= shadowYmin) {
 						final int xintercept = (int) Math.round(sx0 + (shadowYmin - sy0) * alpha);
